add: priotize impossible goal in children goal list
