+++ src/components/GoalsComponents/GoalSublist/GoalSublist.tsx
@@ -21,6 +21,7 @@ import ConfigGoal from "../GoalConfigModal/ConfigGoal";
 import "./GoalSublist.scss";
 import GoalHistory from "./GoalHistory";
 import ArchivedAccordion from "../ArchivedAccordion";
+import { priotizeImpossibleGoals } from "@src/utils/priotizeImpossibleGoals";
 
 export const GoalSublist = () => {
   const { t } = useTranslation();
@@ -36,9 +37,10 @@ export const GoalSublist = () => {
   const [archivedChildren, setArchivedChildren] = useState<GoalItem[]>([]);
   const [showActions, setShowActions] = useState({ open: "root", click: 1 });
 
-  const handleChildrenGoals = (goals: GoalItem[]) => {
-    setChildrenGoals([...goals.filter((goal) => goal.archived === "false")]);
-    setArchivedChildren([...goals.filter((goal) => goal.archived === "true")]);
+  const handleChildrenGoals = async (goals: GoalItem[]) => {
+    const sortedGoals = await priotizeImpossibleGoals(goals);
+    setChildrenGoals([...sortedGoals.filter((goal) => goal.archived === "false")]);
+    setArchivedChildren([...sortedGoals.filter((goal) => goal.archived === "true")]);
   };
 
   useEffect(() => {
@@ -56,7 +58,7 @@ export const GoalSublist = () => {
       <GoalHistory />
       <div className="sublist-content-container">
         <div className="sublist-content">
-          <p className="sublist-title">{t(parentGoal?.title)}</p>
+          <p className="sublist-title">{parentGoal && t(parentGoal?.title)}</p>
           <div className="sublist-list-container">
             {showAddGoal && <ConfigGoal action="Create" goal={createGoalObjectFromTags({})} />}
             <GoalsList
